refactor(Restaurant): HomeEventHandler 메서드 Restaurant로 역할 분리
